code re-structuring
